Set 2 Refactoring Extract class method applied on Channel.java and cretaed new class ModeCLass.java,where some of the methods from Channel.java to ModeClass.java class in order to make it to follow single responsibility principle.
